Memory allocation and merging of free regions now work!

+++ kernel/memory.h
@@ -1,12 +1,14 @@
 #ifndef MEMORY_H_
 #define MEMORY_H_
 
+#include <stddef.h>
+
 void memory_init(void);
 
-void *malloc(unsigned int size);
+void *malloc(size_t size);
 void free(void *p);
 
-void memory_zero(void *p, unsigned int size);
+void memory_zero(void *p, size_t size);
 
 void debug_print_free_regions(void);
